+++ src/components/Typography/Info.jsx
@@ -16,7 +16,8 @@ function Info({ ...props }) {
 }
 
 Info.propTypes = {
-  classes: PropTypes.object.isRequired
+  classes: PropTypes.object.isRequired,
+  children: PropTypes.node
 };
 
 export default withStyles(typographyStyle)(Info);