+++ src/views/Components/Sections/SectionJavascript.jsx
@@ -1,4 +1,6 @@
 import React from "react";
+// nodejs library to set properties for components
+import PropTypes from "prop-types";
 // react plugin for creating date-time-picker
 import Datetime from "react-datetime";
 // @material-ui/core components
@@ -337,5 +339,9 @@ class SectionJavascript extends React.Component {
     );
   }
 }
+
+SectionJavascript.propTypes = {
+  classes: PropTypes.object
+};
 
 export default withStyles(javascriptStyles)(SectionJavascript);