+++ src/views/Components/Sections/SectionLogin.jsx
@@ -1,4 +1,6 @@
 import React from "react";
+// nodejs library to set properties for components
+import PropTypes from "prop-types";
 // @material-ui/core components
 import withStyles from "@material-ui/core/styles/withStyles";
 import InputAdornment from "@material-ui/core/InputAdornment";
@@ -133,5 +135,9 @@ class SectionLogin extends React.Component {
     );
   }
 }
+
+SectionLogin.propTypes = {
+  classes: PropTypes.object
+};
 
 export default withStyles(loginStyle)(SectionLogin);