+++ src/views/Components/Sections/SectionTypography.jsx
@@ -1,5 +1,6 @@
 import React from "react";
-
+// nodejs library to set properties for components
+import PropTypes from "prop-types";
 // @material-ui/core components
 import withStyles from "@material-ui/core/styles/withStyles";
 
@@ -133,7 +134,7 @@ class SectionTypography extends React.Component {
                 <h2>
                   Header with small subtitle
                   <br />
-                  <Small>Use "Small" tag for the headers</Small>
+                  <Small>Use {'"Small"'} tag for the headers</Small>
                 </h2>
               </div>
             </GridContainer>
@@ -198,5 +199,9 @@ class SectionTypography extends React.Component {
     );
   }
 }
+
+SectionTypography.propTypes = {
+  classes: PropTypes.object
+};
 
 export default withStyles(typographyStyle)(SectionTypography);