+++ src/views/LandingPage/LandingPage.jsx
@@ -1,4 +1,6 @@
 import React from "react";
+// nodejs library to set properties for components
+import PropTypes from "prop-types";
 // nodejs library that concatenates classes
 import classNames from "classnames";
 // @material-ui/core components
@@ -48,9 +50,9 @@ class LandingPage extends React.Component {
                 <h1 className={classes.title}>Your Story Starts With Us.</h1>
                 <h4>
                   Every landing page needs a small description after the big
-                  bold title, that's why we added this text here. Add here all
-                  the information that can make you or your product create the
-                  first impression.
+                  bold title, that{"'"}s why we added this text here. Add here
+                  all the information that can make you or your product create
+                  the first impression.
                 </h4>
                 <br />
                 <Button
@@ -79,5 +81,9 @@ class LandingPage extends React.Component {
     );
   }
 }
+
+LandingPage.propTypes = {
+  classes: PropTypes.object
+};
 
 export default withStyles(landingPageStyle)(LandingPage);