+++ src/views/LandingPage/Sections/ProductSection.jsx
@@ -1,4 +1,6 @@
 import React from "react";
+// nodejs library to set properties for components
+import PropTypes from "prop-types";
 // @material-ui/core components
 import withStyles from "@material-ui/core/styles/withStyles";
 
@@ -20,13 +22,13 @@ class ProductSection extends React.Component {
       <div className={classes.section}>
         <GridContainer justify="center">
           <GridItem xs={12} sm={12} md={8}>
-            <h2 className={classes.title}>Let's talk product</h2>
+            <h2 className={classes.title}>Let{"'"}s talk product</h2>
             <h5 className={classes.description}>
               This is the paragraph where you can write more details about your
               product. Keep you user engaged by providing meaningful
               information. Remember that by this time, the user is curious,
-              otherwise he wouldn't scroll to get here. Add a button if you want
-              the user to see more.
+              otherwise he wouldn{"'"}t scroll to get here. Add a button if you
+              want the user to see more.
             </h5>
           </GridItem>
         </GridContainer>
@@ -65,5 +67,9 @@ class ProductSection extends React.Component {
     );
   }
 }
+
+ProductSection.propTypes = {
+  classes: PropTypes.object
+};
 
 export default withStyles(productStyle)(ProductSection);