+++ src/views/ProfilePage/ProfilePage.jsx
@@ -1,4 +1,6 @@
 import React from "react";
+// nodejs library to set properties for components
+import PropTypes from "prop-types";
 // nodejs library that concatenates classes
 import classNames from "classnames";
 // @material-ui/core components
@@ -212,5 +214,9 @@ class ProfilePage extends React.Component {
     );
   }
 }
+
+ProfilePage.propTypes = {
+  classes: PropTypes.object
+};
 
 export default withStyles(profilePageStyle)(ProfilePage);