+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Removed `.env` file, and replaced it with the `jsconfig.json` file
 - Changes caused by running [the prettier command](https://prettier.io/docs/en/install.html) for *.jsx*, *.js*, *.html* and *.css* files
 - Changed all string refs to `React.createRef()`
+- Added types validation in each component
 - Solved https://github.com/creativetimofficial/material-kit-react/issues/52
 - Solved https://github.com/creativetimofficial/ct-material-dashboard-pro-react/issues/149
 ### Major style changes